requirements.txt
pygame-ce>=2.4.0
torch>=2.2.0
numpy>=1.26.0
